+++ cmd/template/cluster/provider/common.go
@@ -1,13 +1,10 @@
 package provider
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"io"
 	"math"
 	"strings"
-	"text/template"
 
 	applicationv1alpha1 "github.com/giantswarm/apiextensions-application/api/v1alpha1"
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
@@ -17,16 +14,9 @@ import (
 	"github.com/giantswarm/micrologger"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/labels"
-	apiyaml "k8s.io/apimachinery/pkg/runtime/serializer/yaml"
-	"k8s.io/client-go/discovery"
-	memory "k8s.io/client-go/discovery/cached"
-	"k8s.io/client-go/dynamic"
-	"k8s.io/client-go/restmapper"
 	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/yaml"
 
 	"github.com/giantswarm/kubectl-gs/v2/pkg/app"
 )
@@ -70,6 +60,29 @@ type GCPConfig struct {
 	FailureDomains    []string
 	ControlPlane      GCPControlPlane
 	MachineDeployment GCPMachineDeployment
+}
+
+type VSphereConfig struct {
+	ControlPlane            VSphereControlPlane
+	CredentialsSecretName   string
+	ImageTemplate           string
+	NetworkName             string
+	Worker                  VSphereMachineTemplate
+	ResourcePool            string
+	ServiceLoadBalancerCIDR string
+}
+
+type VSphereMachineTemplate struct {
+	DiskGiB   int
+	MemoryMiB int
+	NumCPUs   int
+	Replicas  int
+}
+
+type VSphereControlPlane struct {
+	IP         string
+	IPPoolName string
+	VSphereMachineTemplate
 }
 
 type GCPControlPlane struct {
@@ -146,6 +159,7 @@ type ClusterConfig struct {
 	App       AppConfig
 	AWS       AWSConfig
 	Azure     AzureConfig
+	VSphere   VSphereConfig
 	GCP       GCPConfig
 	OpenStack OpenStackConfig
 }
@@ -156,11 +170,6 @@ type OIDC struct {
 	ClientID      string
 	UsernameClaim string
 	GroupsClaim   string
-}
-
-type templateConfig struct {
-	Name string
-	Data string
 }
 
 func newcapiClusterCR(config ClusterConfig, infrastructureRef *corev1.ObjectReference) *capi.Cluster {
@@ -198,95 +207,6 @@ func newcapiClusterCR(config ClusterConfig, infrastructureRef *corev1.ObjectRefe
 	}
 
 	return cluster
-}
-
-func runMutation(ctx context.Context, client k8sclient.Interface, templateData interface{}, templates []templateConfig, output io.Writer) error {
-	var err error
-
-	// Create a DiscoveryClient to get the GVR.
-	dc, err := discovery.NewDiscoveryClientForConfig(client.RESTConfig())
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	// Create a DynamicClient to execute our request.
-	dyn, err := dynamic.NewForConfig(client.RESTConfig())
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	mapper := restmapper.NewDeferredDiscoveryRESTMapper(memory.NewMemCacheClient(dc))
-	for _, t := range templates {
-		// Add separators to make the entire file valid yaml and allow easy appending.
-		_, err = output.Write([]byte("---\n"))
-		if err != nil {
-			return microerror.Mask(err)
-		}
-		te := template.Must(template.New(t.Name).Parse(t.Data))
-		var buf bytes.Buffer
-		// Template from our inputs.
-		err = te.Execute(&buf, templateData)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-
-		// Transform to unstructured.Unstructured.
-		obj := &unstructured.Unstructured{}
-		dec := apiyaml.NewDecodingSerializer(unstructured.UnstructuredJSONScheme)
-		_, gvk, err := dec.Decode(buf.Bytes(), nil, obj)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-		// Mapping needed for the dynamic client.
-		mapping, err := mapper.RESTMapping(gvk.GroupKind(), gvk.Version)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-		// Get namespace information from object.
-		namespace, namespaced, err := unstructured.NestedString(obj.Object, "metadata", "namespace")
-		if err != nil {
-			return microerror.Mask(err)
-		}
-		var defaultedObj *unstructured.Unstructured
-		// Execute our request as a `DryRun` - no resources will be persisted.
-		if namespaced {
-			defaultedObj, err = dyn.Resource(mapping.Resource).Namespace(namespace).Create(ctx, obj, metav1.CreateOptions{DryRun: []string{"All"}})
-			if err != nil {
-				// TODO handle different kinds of errors (e.g. validation) here.
-				return microerror.Mask(err)
-			}
-		} else {
-			defaultedObj, err = dyn.Resource(mapping.Resource).Create(ctx, obj, metav1.CreateOptions{DryRun: []string{"All"}})
-			if err != nil {
-				// TODO handle different kinds of errors (e.g. validation) here.
-				return microerror.Mask(err)
-			}
-		}
-		// Strip `managedFields` for better readability.
-		unstructured.RemoveNestedField(defaultedObj.Object, "metadata", "managedFields")
-		// Strip some metadata fields for better UX.
-		unstructured.RemoveNestedField(defaultedObj.Object, "metadata", "creationTimestamp")
-		unstructured.RemoveNestedField(defaultedObj.Object, "metadata", "generation")
-		unstructured.RemoveNestedField(defaultedObj.Object, "metadata", "selfLink")
-		unstructured.RemoveNestedField(defaultedObj.Object, "metadata", "uid")
-		// Unstructured to JSON.
-		buf = *new(bytes.Buffer)
-		err = dec.Encode(defaultedObj, &buf)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-		// JSON to YAML.
-		mutated, err := yaml.JSONToYAML(buf.Bytes())
-		if err != nil {
-			return microerror.Mask(err)
-		}
-		// Write the yaml to our file.
-		_, err = output.Write(mutated)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-	}
-	return nil
 }
 
 func getLatestVersion(ctx context.Context, ctrlClient client.Client, app, catalog string) (string, error) {